feat: custom hook for closing detailedCard
